Move EOS logic into separate class file
EosResources.java now holds all of the EOS price calcuation logic
